move linked apps related methods to linked_domain app
