+++ corehq/apps/app_manager/tasks.py
@@ -86,12 +86,6 @@ def update_linked_app_and_notify_task(domain, app_id, user_id, email):
     update_linked_app_and_notify(domain, app_id, user_id, email)
 
 
-@task(queue='background_queue')
-def pull_missing_multimedia_for_app_and_notify_task(domain, app_id, email=None):
-    from corehq.apps.app_manager.views.utils import pull_missing_multimedia_for_app_and_notify
-    pull_missing_multimedia_for_app_and_notify(domain, app_id, email)
-
-
 @task
 def load_appcues_template_app(domain, username, app_slug):
     from corehq.apps.app_manager.views.apps import load_app_from_slug
