+++ corehq/apps/linked_domain/tests/test_linked_apps.py
@@ -14,12 +14,17 @@ from corehq.apps.app_manager.models import (
 from corehq.apps.linked_domain.dbaccessors import get_domain_master_link
 from corehq.apps.linked_domain.exceptions import ActionNotPermitted
 from corehq.apps.linked_domain.models import DomainLink, RemoteLinkDetails
-from corehq.apps.linked_domain.remote_accessors import _convert_app_from_remote_linking_source, \
-    _get_missing_multimedia, _fetch_remote_media
+from corehq.apps.linked_domain.remote_accessors import (
+    _convert_app_from_remote_linking_source,
+    fetch_remote_media,
+)
 from corehq.apps.app_manager.tests.util import TestXmlMixin
 from corehq.apps.app_manager.views.utils import overwrite_app, _get_form_ids_by_xmlns, update_linked_app
 from corehq.apps.hqmedia.models import CommCareImage, CommCareMultimedia
-from corehq.apps.linked_domain.util import convert_app_for_remote_linking
+from corehq.apps.linked_domain.util import (
+    convert_app_for_remote_linking,
+    _get_missing_multimedia,
+)
 from io import open
 
 from corehq.util.test_utils import flag_enabled, softer_assert
@@ -323,7 +328,7 @@ class TestRemoteLinkedApps(BaseLinkedAppsTest):
         media_details['media_type'] = 'CommCareMultimedia'
         with patch('corehq.apps.linked_domain.remote_accessors._fetch_remote_media_content') as mock:
             mock.return_value = data
-            _fetch_remote_media('domain', [('case_list_image.jpg', media_details)], remote_details)
+            fetch_remote_media('domain', [('case_list_image.jpg', media_details)], remote_details)
 
         media = CommCareMultimedia.get(media_details['multimedia_id'])
         self.addCleanup(media.delete)
